Move UpnpService out of p2p module
